Move scenario handling into main entrypoint
The factorio CLI argument --map2scenario does not respect settings from map-settings.json, and is not supported for this use any more by the devs. Removing the two alternate entrypoints for scenario handling and moving the ability to launch a new clean map from a scenario into the primary entrypoint.