Remove unneccessary access token from update workflow
It is not necessary to use a PAT to check out a public repository. Because it is defined, the PAT eventually will expire and the action will stop working. By leaving it at default, the workflow generates a new Github token every time it runs which is enough to check out the repo.